fix: code style and enhance user experience
