+++ package.json
@@ -60,6 +60,7 @@
     "react-dom": "^18.0.1",
     "react-dropzone": "^14.2.3",
     "react-error-boundary": "^3.1.4",
+    "react-hooks-global-state": "^2.1.0",
     "react-hot-toast": "^2.4.0",
     "react-hotkeys-hook": "^4.3.7",
     "react-i18next": "^12.2.0",
