+++ src/common/highlight-in-textarea/index.ts
@@ -1,8 +1,7 @@
 import styles from 'inline:./index.css'
 
-/* eslint-disable @typescript-eslint/no-explicit-any */
 interface IConfig {
-    highlight: string | RegExp
+    highlight: string | RegExp | string[]
 }
 
 export class HighlightInTextarea {
@@ -148,7 +147,7 @@ export class HighlightInTextarea {
 
     public handleInput() {
         const input = this.el?.value
-        const ranges = this.getRanges(input, this.highlight)
+        const ranges = this.getRanges(input, this.highlight?.highlight ?? null)
         const unstaggeredRanges = this.removeStaggeredRanges(ranges)
         const boundaries = this.getBoundaries(unstaggeredRanges)
         this.renderMarks(boundaries)
@@ -180,7 +179,7 @@ export class HighlightInTextarea {
     }
 
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    getRanges(input: string | undefined, highlight: IConfig | null): any {
+    getRanges(input: string | undefined, highlight: IConfig['highlight'] | null): any {
         const type = this.getType(highlight)
         switch (type) {
             case 'array':
