+++ src/common/hooks/global.ts
@@ -1,0 +1,4 @@
+import { createGlobalState } from 'react-hooks-global-state'
+
+const initialState = { collectedWordTotal: 0 }
+export const { useGlobalState } = createGlobalState(initialState)
