+++ src/content_script/PopupCard.tsx
@@ -17,12 +17,12 @@ import { translate, TranslateMode } from './translate'
 import { Select, Value, Option } from 'baseui-sd/select'
 import { CopyToClipboard } from 'react-copy-to-clipboard'
 import { RxCopy, RxEraser, RxReload, RxSpeakerLoud } from 'react-icons/rx'
-import { calculateMaxXY, exportToCsv, LocalDB, queryPopupCardElement, VocabularyItem } from './utils'
+import { calculateMaxXY, queryPopupCardElement } from './utils'
 import { clsx } from 'clsx'
 import { Button } from 'baseui-sd/button'
 import { ErrorBoundary } from 'react-error-boundary'
 import { ErrorFallback } from '../components/ErrorFallback'
-import { defaultAPIURL, isDesktopApp, isTauri } from '../common/utils'
+import { defaultAPIURL, exportToCsv, isDesktopApp, isTauri } from '../common/utils'
 import { Settings } from '../popup/Settings'
 import { documentPadding } from './consts'
 import Dropzone from 'react-dropzone'
@@ -41,7 +41,10 @@ import { useTheme } from '../common/hooks/useTheme'
 import { speak } from '../common/tts'
 import { Tooltip } from '../components/Tooltip'
 import { useSettings } from '../common/hooks/useSettings'
-import Vocabulary from './vocalulary'
+import Vocabulary from './Vocabulary'
+import { LocalDB, VocabularyItem } from '../common/db'
+import { useCollectedWordTotal } from '../common/hooks/useCollectedWordTotal'
+import { Modal } from 'baseui-sd/modal'
 
 const cache = new LRUCache({
     max: 500,
@@ -394,7 +397,7 @@ export function PopupCard(props: IPopupCardProps) {
     const [selectedWord, setSelectedWord] = useState('')
     const [vocabularyType, setVocabularyType] = useState<'hide' | 'vocabulary' | 'article'>('hide')
     const highlightRef = useRef<HighlightInTextarea | null>(null)
-    const [showMoreBtn, setShowMoreBtn] = useState(false)
+    const [showWordbookButtons, setShowWordbookButtons] = useState(false)
     const { t, i18n } = useTranslation()
     const { settings } = useSettings()
     useEffect(() => {
@@ -427,13 +430,19 @@ export function PopupCard(props: IPopupCardProps) {
         }
     }, [props.autoFocus])
 
+    const [highlightWords, setHighlightWords] = useState<string[]>([])
+
     useEffect(() => {
         if (!highlightRef.current?.highlight) {
             return
         }
-        highlightRef.current.highlight.highlight = selectedWord
+        if (selectedWord) {
+            highlightRef.current.highlight.highlight = [selectedWord]
+        } else {
+            highlightRef.current.highlight.highlight = [...highlightWords]
+        }
         highlightRef.current.handleInput()
-    }, [selectedWord])
+    }, [selectedWord, highlightWords])
 
     const [translateMode, setTranslateMode] = useState<TranslateMode | ''>('')
     useEffect(() => {
@@ -458,13 +467,15 @@ export function PopupCard(props: IPopupCardProps) {
             isCompositing.current = false
         }
         const onMouseUp = () => {
+            if (editor.selectionStart === 0 && editor.selectionEnd === editor.value.length) {
+                setSelectedWord('')
+                return
+            }
             const selectedWord_ = editor.value.substring(editor.selectionStart, editor.selectionEnd).trim()
             setSelectedWord(selectedWord_)
+            setHighlightWords([])
         }
-        const onBlur = () => {
-            const selectedWord_ = editor.value.substring(editor.selectionStart, editor.selectionEnd).trim()
-            setSelectedWord(selectedWord_)
-        }
+        const onBlur = onMouseUp
 
         editor.addEventListener('compositionstart', onCompositionStart)
         editor.addEventListener('compositionend', onCompositionEnd)
@@ -488,30 +499,29 @@ export function PopupCard(props: IPopupCardProps) {
     const [originalText, setOriginalText] = useState(props.text)
     const [translatedText, setTranslatedText] = useState('')
     const [translatedLines, setTranslatedLines] = useState<string[]>([])
-    const [wordMode, setWordMode] = useState<boolean>(false)
-    const [isCollectd, setIsCollected] = useState(false)
-    const [collectTotal, setCollectTotal] = useState(0)
-    const checkCollection = useCallback(async () => {
+    const [isWordMode, setIsWordMode] = useState(false)
+    const [isCollectedWord, setIsCollectedWord] = useState(false)
+    const checkWordCollection = useCallback(async () => {
         try {
-            const arr = await LocalDB.vocabulary.where('word').equals(editableText).toArray()
+            const arr = await LocalDB.vocabulary.where('word').equals(editableText.trim()).toArray()
             if (arr.length > 0) {
                 await LocalDB.vocabulary.put({
                     ...arr[0],
-                    count: arr[0].count + 1,
+                    reviewCount: arr[0].reviewCount + 1,
                 })
-                setIsCollected(true)
+                setIsCollectedWord(true)
             } else {
-                setIsCollected(false)
+                setIsCollectedWord(false)
             }
         } catch (e) {
             console.error(e)
         }
     }, [editableText])
     useEffect(() => {
-        if (wordMode && !isLoading) {
-            checkCollection()
+        if (isWordMode && !isLoading) {
+            checkWordCollection()
         }
-    }, [wordMode, editableText, isLoading])
+    }, [isWordMode, editableText, isLoading])
     useEffect(() => {
         setTranslatedLines(translatedText.split('\n'))
     }, [translatedText])
@@ -568,17 +578,10 @@ export function PopupCard(props: IPopupCardProps) {
 
     const scrollYRef = useRef<number>(0)
 
+    const { collectedWordTotal, setCollectedWordTotal } = useCollectedWordTotal()
+
     // Reposition the popup card to prevent it from extending beyond the screen.
     useEffect(() => {
-        const initCollectionTotal = async () => {
-            try {
-                const total = await LocalDB.vocabulary.count()
-                setCollectTotal(total)
-            } catch (e) {
-                console.error(e)
-            }
-        }
-        initCollectionTotal()
         const calculateTranslatedContentMaxHeight = (): number => {
             const { innerHeight } = window
             const headerHeight = headerRef.current?.offsetHeight || 0
@@ -716,7 +719,7 @@ export function PopupCard(props: IPopupCardProps) {
 
     const translateText = useCallback(
         async (text: string, selectedWord: string, signal: AbortSignal) => {
-            setShowMoreBtn(false)
+            setShowWordbookButtons(false)
             if (!text || !detectFrom || !detectTo || !translateMode) {
                 return
             }
@@ -772,7 +775,7 @@ export function PopupCard(props: IPopupCardProps) {
                         if (message.role) {
                             return
                         }
-                        setWordMode(message.wordMode)
+                        setIsWordMode(message.isWordMode)
                         setTranslatedText((translatedText) => {
                             return translatedText + message.content
                         })
@@ -786,6 +789,9 @@ export function PopupCard(props: IPopupCardProps) {
                                 ['”', '"', '」'].indexOf(translatedText[translatedText.length - 1]) >= 0
                             ) {
                                 result = translatedText.slice(0, -1)
+                            }
+                            if (result && ['“', '"', '「'].indexOf(result[0]) >= 0) {
+                                result = result.slice(1)
                             }
                             cache.set(cachedKey, result)
                             return result
@@ -950,20 +956,21 @@ export function PopupCard(props: IPopupCardProps) {
 
     const onWordCollection = async () => {
         try {
-            if (isCollectd) {
+            if (isCollectedWord) {
                 const wordInfo = await LocalDB.vocabulary.get(editableText)
                 await LocalDB.vocabulary.delete(wordInfo?.word ?? '')
-                setIsCollected(false)
-                setCollectTotal((t) => t - 1)
+                setIsCollectedWord(false)
+                setCollectedWordTotal((t) => t - 1)
             } else {
                 await LocalDB.vocabulary.put({
                     word: editableText,
-                    count: 1,
-                    description: translatedText.substr(editableText.length + 1), // separate string after first '\n'
-                    updateAt: new Date().valueOf().toString(),
+                    reviewCount: 1,
+                    description: translatedText.slice(editableText.length + 1), // separate string after first '\n'
+                    updatedAt: new Date().valueOf().toString(),
+                    createdAt: new Date().valueOf().toString(),
                 })
-                setIsCollected(true)
-                setCollectTotal((t) => t + 1)
+                setIsCollectedWord(true)
+                setCollectedWordTotal((t) => t + 1)
             }
         } catch (e) {
             console.error(e)
@@ -972,11 +979,10 @@ export function PopupCard(props: IPopupCardProps) {
 
     const onCsvExport = async () => {
         try {
-            const arr = await LocalDB.vocabulary.toArray()
-            console.log('arr==', arr)
-            await exportToCsv<VocabularyItem>(`openai-translator-collection-${new Date().valueOf()}`, arr)
+            const words = await LocalDB.vocabulary.toArray()
+            await exportToCsv<VocabularyItem>(`openai-translator-collection-${new Date().valueOf()}`, words)
             if (isDesktopApp()) {
-                toast(t('Csv saved on Desktop'), {
+                toast(t('csv file saved on Desktop'), {
                     duration: 5000,
                     icon: '👏',
                 })
@@ -1337,13 +1343,13 @@ export function PopupCard(props: IPopupCardProps) {
                                                         </Dropzone>
                                                     </div>
                                                 </Tooltip>
-                                                {!!collectTotal && (
+                                                {collectedWordTotal > 0 && (
                                                     <StatefulTooltip
                                                         content={
                                                             <Trans
                                                                 i18nKey='words are collected'
                                                                 values={{
-                                                                    collectTotal,
+                                                                    collectTotal: collectedWordTotal,
                                                                 }}
                                                             />
                                                         }
@@ -1352,13 +1358,13 @@ export function PopupCard(props: IPopupCardProps) {
                                                     >
                                                         <div
                                                             className={styles.actionButton}
-                                                            onClick={() => setShowMoreBtn((e) => !e)}
+                                                            onClick={() => setShowWordbookButtons((e) => !e)}
                                                         >
                                                             <AiOutlineFileSync size={13} />
                                                         </div>
                                                     </StatefulTooltip>
                                                 )}
-                                                {showMoreBtn && (
+                                                {showWordbookButtons && collectedWordTotal > 0 && (
                                                     <>
                                                         <StatefulTooltip
                                                             content={t('Collection Review')}
@@ -1373,7 +1379,7 @@ export function PopupCard(props: IPopupCardProps) {
                                                             </div>
                                                         </StatefulTooltip>
                                                         <StatefulTooltip
-                                                            content={t('Export csv of your collection')}
+                                                            content={t('Export your collection as a csv file')}
                                                             showArrow
                                                             placement='top'
                                                         >
@@ -1494,7 +1500,7 @@ export function PopupCard(props: IPopupCardProps) {
                                                             {translatedLines.map((line, i) => {
                                                                 return (
                                                                     <p className={styles.paragraph} key={`p-${i}`}>
-                                                                        {wordMode && i == 0 ? (
+                                                                        {isWordMode && i == 0 ? (
                                                                             <div
                                                                                 style={{
                                                                                     display: 'flex',
@@ -1506,7 +1512,7 @@ export function PopupCard(props: IPopupCardProps) {
                                                                                 {!isLoading && (
                                                                                     <StatefulTooltip
                                                                                         content={
-                                                                                            isCollectd
+                                                                                            isCollectedWord
                                                                                                 ? t(
                                                                                                       'Remove from collection'
                                                                                                   )
@@ -1521,7 +1527,7 @@ export function PopupCard(props: IPopupCardProps) {
                                                                                             }
                                                                                             onClick={onWordCollection}
                                                                                         >
-                                                                                            {isCollectd ? (
+                                                                                            {isCollectedWord ? (
                                                                                                 <MdGrade size={15} />
                                                                                             ) : (
                                                                                                 <MdOutlineGrade
@@ -1612,15 +1618,35 @@ export function PopupCard(props: IPopupCardProps) {
                                 </Tooltip>
                             </div>
                         )}
-                        {vocabularyType !== 'hide' && (
-                            <div className={styles.vocabulary} onClick={() => setVocabularyType('hide')}>
-                                <Vocabulary
-                                    onCancel={() => setVocabularyType('hide')}
-                                    type={vocabularyType}
-                                    engine={props.engine}
-                                />
-                            </div>
-                        )}
+                        <Modal
+                            isOpen={vocabularyType !== 'hide'}
+                            onClose={() => setVocabularyType('hide')}
+                            closeable
+                            overrides={{
+                                Close: {
+                                    style: {
+                                        display: 'none',
+                                    },
+                                },
+                            }}
+                            size='auto'
+                            autoFocus
+                            animate
+                            role='dialog'
+                        >
+                            <Vocabulary
+                                onCancel={() => setVocabularyType('hide')}
+                                onInsert={(content, highlightWords) => {
+                                    setEditableText(content)
+                                    setOriginalText(content)
+                                    setHighlightWords(highlightWords)
+                                    setSelectedWord('')
+                                    setTranslateMode('translate')
+                                    setVocabularyType('hide')
+                                }}
+                                type={vocabularyType as 'vocabulary' | 'article'}
+                            />
+                        </Modal>
                         <Toaster />
                     </div>
                 </BaseProvider>
