+++ src/content_script/Vocabulary.tsx
@@ -1,0 +1,481 @@
+import { useCallback, useEffect, useMemo, useState, useRef } from 'react'
+import { Button } from 'baseui-sd/button'
+import { useTheme } from '../common/hooks/useTheme'
+import { createUseStyles } from 'react-jss'
+import { StatefulTooltip } from 'baseui-sd/tooltip'
+import { IThemedStyleProps } from '../common/types'
+import { MdOutlineGrade, MdGrade } from 'react-icons/md'
+import { useTranslation } from 'react-i18next'
+import { FaDice } from 'react-icons/fa'
+import { AiOutlineCloseCircle } from 'react-icons/ai'
+import { Select } from 'baseui-sd/select'
+import { FcIdea } from 'react-icons/fc'
+import { toast } from 'react-hot-toast'
+import { translate } from './translate'
+import CopyToClipboard from 'react-copy-to-clipboard'
+import { RxCopy } from 'react-icons/rx'
+import { format } from 'date-fns'
+import { LocalDB, VocabularyItem } from '../common/db'
+import { useCollectedWordTotal } from '../common/hooks/useCollectedWordTotal'
+import { RiPictureInPictureExitLine } from 'react-icons/ri'
+import { Tooltip } from '../components/Tooltip'
+import { isDesktopApp } from '../common/utils'
+
+const RANDOM_SIZE = 10
+const MAX_WORDS = 50
+
+const useStyles = createUseStyles({
+    'container': (props: IThemedStyleProps) => ({
+        position: 'relative',
+        minWidth: '600px',
+        maxWidth: props.isDesktopApp ? undefined : '800px',
+        minHeight: '480px',
+        background: props.theme.colors.backgroundPrimary,
+        borderRadius: '6px',
+        boxSizing: 'border-box',
+        padding: '10px',
+        display: 'flex',
+        flexDirection: 'column',
+    }),
+    'closeBtn': {
+        'cursor': 'pointer',
+        'zIndex': '1',
+        '&:active': {
+            opacity: 0.6,
+        },
+    },
+    'list': {
+        position: 'relative',
+        display: 'flex',
+        gap: '6px',
+        flexWrap: 'wrap',
+        padding: '16px 0',
+        borderBottom: '1px solid rgba(0, 0, 0, 0.08)',
+    },
+    'display': (props: IThemedStyleProps) => ({
+        display: 'flex',
+        padding: '16px 10px 10px',
+        flexDirection: 'column',
+        overflowY: 'auto',
+        overflowX: 'hidden',
+        color: props.themeType === 'dark' ? props.theme.colors.contentSecondary : props.theme.colors.contentPrimary,
+    }),
+    'articleDisplay': {
+        marginTop: '16px',
+        display: 'flex',
+        padding: '0 10px 10px',
+        flexDirection: 'column',
+        overflowY: 'auto',
+        overflowX: 'hidden',
+    },
+    'diceArea': {
+        position: 'absolute',
+        bottom: 0,
+        left: '50%',
+        transform: 'translateX(-50%) translateY(50%)',
+        fontSize: '10px',
+        borderRadius: '4px',
+        padding: '0 12px',
+        background: '#FFF',
+    },
+    'diceIcon': {
+        'cursor': 'pointer',
+        '&:active': {
+            opacity: 0.6,
+        },
+    },
+    'actionButtonsContainer': {
+        display: 'flex',
+        flexDirection: 'row',
+        alignItems: 'center',
+        gap: '12px',
+    },
+    'actionButton': (props: IThemedStyleProps) => ({
+        color: props.theme.colors.contentSecondary,
+        cursor: 'pointer',
+        display: 'flex',
+        paddingTop: '6px',
+        paddingBottom: '6px',
+    }),
+    'select': {
+        display: 'flex',
+        flexDirection: 'row',
+        alignItems: 'center',
+        width: '100%',
+        gap: '8px',
+    },
+    'actionStr': (props: IThemedStyleProps) => ({
+        position: 'absolute',
+        display: 'flex',
+        flexDirection: 'row',
+        alignItems: 'center',
+        gap: '6px',
+        bottom: '0',
+        left: '50%',
+        transform: 'translateX(-50%) translateY(50%)',
+        fontSize: '10px',
+        padding: '2px 12px',
+        borderRadius: '4px',
+        background: props.theme.colors.backgroundTertiary,
+        color: props.theme.colors.contentSecondary,
+    }),
+    'error': {
+        background: '#f8d7da',
+    },
+    'writing': {
+        'marginLeft': '3px',
+        'width': '10px',
+        '&::after': {
+            content: '"✍️"',
+            animation: '$writing 1.3s infinite',
+        },
+    },
+    '@keyframes writing': {
+        '50%': {
+            marginLeft: '-3px',
+            marginBottom: '-3px',
+        },
+    },
+})
+
+interface IVocabularyProps {
+    type: 'vocabulary' | 'article'
+    onCancel: () => void
+    onInsert: (content: string, highlightWords: string[]) => void
+}
+
+const Vocabulary = (props: IVocabularyProps) => {
+    const { theme, themeType } = useTheme()
+    const styles = useStyles({ theme, themeType, isDesktopApp: isDesktopApp() })
+
+    const [words, setWords] = useState<VocabularyItem[]>([])
+    const [selectedWord, setSelectedWord] = useState<VocabularyItem>()
+    const [isCollectedWord, setIsCollectedWord] = useState(false)
+    const { t } = useTranslation()
+    const controlRef = useRef(new AbortController())
+    const articleOptions = [
+        {
+            id: 'story',
+            prompt: 'an insteresting story',
+            label: t('An insteresting story'),
+        },
+        {
+            id: 'newsletter',
+            prompt: 'a political newsletter',
+            label: t('A political newsletter'),
+        },
+        {
+            id: 'sports',
+            prompt: 'a sports bulletin',
+            label: t('A sports bulletin'),
+        },
+        {
+            id: 'lyric',
+            prompt: 'a catchy lyric',
+            label: t('A catchy lyric'),
+        },
+        {
+            id: 'poem',
+            prompt: 'a smooth poem',
+            label: t('A smooth poem'),
+        },
+    ]
+    const [articleType, setArticleType] = useState<string>(articleOptions[0].id)
+    const [isLoading, setIsLoading] = useState<boolean>(false)
+    const [article, setArticle] = useState<string>('')
+    const articleTxt = useRef<string>('')
+    const descriptionLines = useMemo(() => selectedWord?.description.split('\n') ?? [], [selectedWord?.description])
+    const articleUsedWord = useRef<string[]>()
+    const { collectedWordTotal, setCollectedWordTotal } = useCollectedWordTotal()
+
+    const checkCollection = useCallback(async () => {
+        try {
+            const count = await LocalDB.vocabulary
+                .where('word')
+                .equals(selectedWord?.word ?? '')
+                .count()
+            setIsCollectedWord(count > 0)
+        } catch (e) {
+            console.error(e)
+        }
+    }, [selectedWord?.word])
+
+    useEffect(() => {
+        checkCollection()
+    }, [selectedWord?.word])
+
+    const onRandomWords = async () => {
+        try {
+            let randomWords: VocabularyItem[] = []
+            setSelectedWord(undefined)
+            if (collectedWordTotal <= RANDOM_SIZE) {
+                randomWords = await LocalDB.vocabulary.toArray()
+            } else {
+                const idxSeen: Set<number> = new Set([])
+                while (idxSeen.size < RANDOM_SIZE) {
+                    // eslint-disable-next-line no-constant-condition
+                    while (true) {
+                        const idx = Math.floor(collectedWordTotal * Math.random())
+                        if (idxSeen.has(idx)) {
+                            continue
+                        }
+                        idxSeen.add(idx)
+                        const words_ = await LocalDB.vocabulary.offset(idx).limit(1).toArray()
+                        randomWords.push(words_[0])
+                        break
+                    }
+                }
+            }
+            setWords(randomWords)
+        } catch (e) {
+            console.error(e)
+        }
+    }
+
+    useEffect(() => {
+        onRandomWords()
+    }, [])
+
+    const onWordCollection = async () => {
+        try {
+            if (isCollectedWord) {
+                const wordInfo = await LocalDB.vocabulary.get(selectedWord?.word ?? '')
+                await LocalDB.vocabulary.delete(wordInfo?.word ?? '')
+                setIsCollectedWord(false)
+                setCollectedWordTotal((prev) => prev - 1)
+            } else {
+                const wordInfo = words.filter((item) => item.word === selectedWord?.word)
+                await LocalDB.vocabulary.put({
+                    ...wordInfo[0],
+                })
+                setIsCollectedWord(true)
+            }
+        } catch (e) {
+            console.error(e)
+        }
+    }
+
+    const onGenerageArticle = async () => {
+        if (!articleType) {
+            toast(t('No article type selected'), {
+                duration: 3000,
+                icon: '😰',
+            })
+            return
+        }
+        setArticle('')
+        const prompt = articleOptions.find((item) => item.id == articleType)?.prompt
+        setIsLoading(true)
+        controlRef.current.abort()
+        controlRef.current = new AbortController()
+        const { signal } = controlRef.current
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        const frequentWords = await (LocalDB.vocabulary.orderBy('reviewCount') as any).desc().limit(MAX_WORDS).toArray()
+        const frequentWordsArr = frequentWords.map((item: VocabularyItem) => item.word)
+        articleUsedWord.current = [...frequentWordsArr]
+        articleTxt.current = ''
+        const str = frequentWordsArr.join(',')
+        try {
+            await translate({
+                mode: 'big-bang',
+                signal,
+                text: str,
+                selectedWord: '',
+                detectFrom: '',
+                detectTo: '',
+                articlePrompt: prompt,
+                onMessage: (message) => {
+                    if (message.role) {
+                        return
+                    }
+                    setArticle((e) => {
+                        articleTxt.current += message.content
+                        if (articleUsedWord.current?.includes(message.content.toLowerCase().trim())) {
+                            return e + `<b style="color: #f40;">${message.content}</b>`
+                        } else {
+                            return e + message.content
+                        }
+                    })
+                },
+                onFinish: () => {
+                    setIsLoading(false)
+                },
+                onError: (error) => {
+                    setIsLoading(false)
+                    toast(error, {
+                        duration: 3000,
+                        icon: '😰',
+                    })
+                },
+            })
+        } finally {
+            setIsLoading(false)
+        }
+    }
+
+    const DangerArticle = useMemo(
+        () => (article ? '<div>' + article.replace(/\n/g, '<br/>').replace(/\r/g, '<br//>') + '</div>' : ''),
+        [article]
+    )
+
+    return (
+        <div
+            className={styles.container}
+            onClick={(e) => {
+                e.stopPropagation()
+                e.preventDefault()
+            }}
+        >
+            <div
+                style={{
+                    display: 'flex',
+                    flexDirection: 'row',
+                    alignItems: 'center',
+                    color: themeType === 'dark' ? theme.colors.contentSecondary : theme.colors.contentPrimary,
+                }}
+            >
+                <div
+                    style={{
+                        fontWeight: '600',
+                    }}
+                >
+                    {t(props.type === 'article' ? 'generate article' : props.type)}
+                </div>
+                <div style={{ flex: 1 }}></div>
+                <div className={styles.closeBtn} onClick={props.onCancel}>
+                    <AiOutlineCloseCircle fontSize={13} />
+                </div>
+            </div>
+            <div className={styles.list}>
+                {props.type == 'vocabulary' && (
+                    <>
+                        {collectedWordTotal > 0
+                            ? words.map((item, index) => (
+                                  <Button
+                                      key={index}
+                                      size='mini'
+                                      kind={selectedWord?.word == item.word ? 'primary' : 'secondary'}
+                                      onClick={() => setSelectedWord(item)}
+                                  >
+                                      {item.word}
+                                  </Button>
+                              ))
+                            : 'no words'}
+                    </>
+                )}
+                {props.type == 'article' && (
+                    <div className={styles.select}>
+                        <Select
+                            size='mini'
+                            clearable={false}
+                            options={articleOptions}
+                            value={[{ id: articleType }]}
+                            onChange={({ value }) => {
+                                controlRef.current.abort()
+                                setArticle('')
+                                setArticleType(`${value[0].id ?? ''}`)
+                            }}
+                        />
+                        <StatefulTooltip content='Big Bang' placement='bottom' showArrow>
+                            <div className={styles.actionButton} onClick={onGenerageArticle}>
+                                <Button size='mini' kind={'secondary'}>
+                                    <FcIdea size={20} />
+                                </Button>
+                            </div>
+                        </StatefulTooltip>
+                    </div>
+                )}
+                {props.type == 'vocabulary' && (
+                    <StatefulTooltip content={t('Random Change')} placement='bottom' showArrow>
+                        <div className={styles.diceArea}>
+                            <FaDice fontSize={20} className={styles.diceIcon} onClick={onRandomWords} />
+                        </div>
+                    </StatefulTooltip>
+                )}
+                {props.type == 'article' && isLoading && (
+                    <div className={styles.actionStr}>
+                        {isLoading && (
+                            <>
+                                <div>{t('is writing')}</div>
+                                <span className={styles.writing} key={'1'} />
+                            </>
+                        )}
+                    </div>
+                )}
+            </div>
+            {props.type == 'vocabulary' && (
+                <div className={styles.display}>
+                    {selectedWord?.word && (
+                        <div
+                            style={{
+                                display: 'flex',
+                                alignItems: 'center',
+                                gap: '5px',
+                            }}
+                        >
+                            {selectedWord.word}
+                            <StatefulTooltip
+                                content={isCollectedWord ? t('Remove from collection') : t('Add to collection')}
+                                showArrow
+                                placement='right'
+                            >
+                                <div className={styles.actionButton} onClick={onWordCollection}>
+                                    {isCollectedWord ? <MdGrade size={15} /> : <MdOutlineGrade size={15} />}
+                                </div>
+                            </StatefulTooltip>
+                        </div>
+                    )}
+                    {descriptionLines.length > 0 && descriptionLines.map((line, idx) => <p key={idx}>{line}</p>)}
+                    {selectedWord?.reviewCount && <p>{`[${t('review count')}] ${selectedWord?.reviewCount}`}</p>}
+                    {selectedWord?.updatedAt && (
+                        <p>{`[${t('last review')}] ${format(+selectedWord?.updatedAt, 'yyyy-MM-dd HH:mm:ss')}`}</p>
+                    )}
+                </div>
+            )}
+            {props.type == 'article' && (
+                <div
+                    className={`${styles.display} ${styles.articleDisplay}`}
+                    dangerouslySetInnerHTML={{ __html: DangerArticle }}
+                ></div>
+            )}
+            {props.type == 'article' && !isLoading && article.length > 0 && (
+                <div
+                    className={styles.actionButtonsContainer}
+                    style={{
+                        marginLeft: 'auto',
+                        marginTop: '10px',
+                    }}
+                >
+                    <Tooltip content={t('insert to editor')}>
+                        <div
+                            className={styles.actionButton}
+                            onClick={() => props.onInsert(articleTxt.current, articleUsedWord.current ?? [])}
+                        >
+                            <RiPictureInPictureExitLine size={13} />
+                        </div>
+                    </Tooltip>
+                    <Tooltip content={t('Copy to clipboard')} showArrow placement='left'>
+                        <div>
+                            <CopyToClipboard
+                                text={articleTxt.current}
+                                onCopy={() => {
+                                    toast(t('Copy to clipboard'), {
+                                        duration: 3000,
+                                        icon: '👏',
+                                    })
+                                }}
+                                options={{ format: 'text/plain' }}
+                            >
+                                <div className={styles.actionButton}>
+                                    <RxCopy size={13} />
+                                </div>
+                            </CopyToClipboard>
+                        </div>
+                    </Tooltip>
+                </div>
+            )}
+        </div>
+    )
+}
+
+export default Vocabulary
