+++ src/content_script/utils.ts
@@ -1,10 +1,8 @@
 import { createParser } from 'eventsource-parser'
 import { backgroundFetch } from '../common/background-fetch'
 import { userscriptFetch } from '../common/userscript-polyfill'
-import { isFirefox, isUserscript, isDesktopApp } from '../common/utils'
+import { isFirefox, isUserscript } from '../common/utils'
 import { containerID, documentPadding, popupCardID, popupThumbID, zIndex } from './consts'
-import Dexie, { Table } from 'dexie'
-import { writeTextFile, BaseDirectory } from '@tauri-apps/api/fs'
 
 function attachEventsToContainer($container: HTMLElement) {
     $container.addEventListener('mousedown', (event) => {
@@ -108,70 +106,4 @@ export function calculateMaxXY($popupCard: HTMLElement): number[] {
     const maxX = scrollLeft + innerWidth - width - documentPadding
     const maxY = scrollTop + innerHeight - height - documentPadding
     return [maxX, maxY]
-}
-
-export interface VocabularyItem {
-    word: string
-    count: number
-    description: string
-    updateAt: string
-    [prop: string]: string | number
-}
-
-export class MySubClassedDexie extends Dexie {
-    vocabulary!: Table<VocabularyItem>
-
-    constructor() {
-        super('collection')
-        this.version(1).stores({
-            vocabulary: 'word,count, description , updateAt',
-        })
-    }
-}
-// init indexdb
-export const LocalDB = new MySubClassedDexie()
-
-// js to csv
-export async function exportToCsv<T extends Record<string, string | number>>(filename: string, rows: T[]) {
-    if (!rows.length) return
-    filename += '.csv'
-    const columns = Object.keys(rows[0])
-    let csvFile = ''
-    for (const key of columns) {
-        csvFile += key + ','
-    }
-    csvFile += '\r\n'
-    const processRow = function (row: T) {
-        let s = ''
-        for (const key of columns) {
-            if (key == 'updateAt') {
-                s += '\t' + `${row[key]}` + ','
-            } else {
-                s += '"' + `${row[key]}` + '"' + ','
-            }
-        }
-        return s + '\r\n'
-    }
-
-    for (let i = 0; i < rows.length; i++) {
-        csvFile += processRow(rows[i])
-    }
-
-    if (isDesktopApp()) {
-        try {
-            return await writeTextFile(filename, csvFile, { dir: BaseDirectory.Desktop })
-        } catch (e) {
-            console.error(e)
-        }
-    } else {
-        const link = document.createElement('a')
-        if (link.download !== undefined) {
-            link.setAttribute('href', 'data:text/csv;charset=utf-8,ufeff' + encodeURIComponent(csvFile))
-            link.setAttribute('download', filename)
-            // link.style.visibility = 'hidden'
-            document.body.appendChild(link)
-            link.click()
-            document.body.removeChild(link)
-        }
-    }
 }
