+++ src/i18n/locales/en/translation.json
@@ -42,8 +42,8 @@
   "Country Not Detected": "We were unable to check if your IP address is in a <0>supported region</0> of OpenAI. Please check your Internet connection, and ensure that you are accessing the API in a supported region of OpenAI, or your account may get banned regardless of your GPT Plus subscription status or any remaining account balance.",
   "words are collected": "{{collectTotal}} words are collected",
   "Collection Review": "Collection Review",
-  "Export csv of your collection": "Export csv of your collection",
-  "Csv saved on Desktop": "Csv saved on Desktop",
+  "Export your collection as a csv file": "Export your collection as a csv file",
+  "csv file saved on Desktop": "csv file saved on Desktop",
   "Remove from collection": "Remove from collection",
   "Add to collection": "Add to collection",
   "Random Change": "Random Change",
@@ -56,5 +56,8 @@
   "A smooth poem": "A smooth poem",
   "No article type selected": "No article type selected",
   "review count": "Review Count",
-  "last review": "Last Review Time"
+  "last review": "Last Review Time",
+  "vocabulary": "Vocabulary",
+  "generate article": "Generate Article",
+  "insert to editor": "Insert to Editor"
 }
