+++ src/i18n/locales/th/translation.json
@@ -41,8 +41,8 @@
   "Country Not Detected": "เราไม่สามารถตรวจสอบได้ว่าที่อยู่ IP ของคุณอยู่ใน<0>พื้นที่ที่รองรับ</0> ของ OpenAI โปรดตรวจสอบการเชื่อมต่ออินเทอร์เน็ตของคุณและตรวจสอบว่าคุณเข้าถึง API ในพื้นที่ที่รองรับของ OpenAI หรือไม่ หากไม่เช่นนั้นบัญชีของคุณอาจถูกแบนโดย OpenAI โดยไม่คำนึงถึงสถานะการสมัครสมาชิก GPT Plus หรือยอดเงินคงเหลือใด ๆ",
   "words are collected": "รวบรวมคำ",
   "Collection Review": "รวบรวมการตรวจสอบ",
-  "Export csv of your collection": "ส่งออก csv ของคอลเลกชัน",
-  "Csv saved on Desktop": "Csv ที่บันทึกไว้บนเดสก์ท็อป",
+  "Export your collection as a csv file": "ส่งออกคอลเลกชันของคุณเป็นไฟล์ csv",
+  "csv file saved on Desktop": "ไฟล์ csv บันทึกไว้ที่ Desktop",
   "Remove from collection": "ลบออกจากคอลเลกชัน",
   "Add to collection": "เพิ่มไปยังคอลเลกชัน",
   "Random Change": "การเปลี่ยนแปลงแบบสุ่ม",
@@ -55,5 +55,8 @@
   "A smooth poem": "บทกวีที่ไหลลื่น",
   "No article type selected":"ยังไม่ได้เลือกประเภทบทความ",
   "review count": "จำนวนการสอบถาม",
-  "last review": "เวลาสอบถามล่าสุด"
+  "last review": "เวลาสอบถามล่าสุด",
+  "vocabulary": "คำศัพท์",
+  "generate article": "สร้างบทความ",
+  "insert to editor": "แทรกลงในตัวแก้ไข"
 }
