+++ src/i18n/locales/zh-Hans/translation.json
@@ -42,8 +42,8 @@
   "Country Not Detected": "我们无法检查您的 IP 地址是否位于 OpenAI 的<0>支持的地区</0>，可能是由于您的网络环境无法访问 OpenAI。请检查您的网络连接，并确保您在支持的位置访问 API，否则即使订阅了 ChatGPT Plus 或账号内仍有余额，您的账户也会被 OpenAI 封禁。",
   "words are collected": "已收藏 {{collectTotal}} 个单词",
   "Collection Review": "单词复习",
-  "Export csv of your collection": "csv导出",
-  "Csv saved on Desktop": "Csv已保存到桌面",
+  "Export your collection as a csv file": "导出收藏为 csv 文件",
+  "csv file saved on Desktop": "csv 文件已保存在桌面",
   "Remove from collection": "移出收藏",
   "Add to collection": "加入收藏",
   "Random Change": "下一批",
@@ -56,5 +56,8 @@
   "A smooth poem": "一首流畅的诗",
   "No article type selected":"未选中文章类型",
   "review count": "查询次数",
-  "last review": "最近查询时间"
+  "last review": "最近查询时间",
+  "vocabulary": "词汇表",
+  "generate article": "生成文章",
+  "insert to editor": "插入编辑器"
 }
